+++ tonic-web/src/lib.rs
@@ -88,13 +88,15 @@
 #![doc(issue_tracker_base_url = "https://github.com/hyperium/tonic/issues/")]
 
 pub use config::Config;
+pub use layer::GrpcWebLayer;
+pub use service::GrpcWeb;
 
 mod call;
 mod config;
 mod cors;
+mod layer;
 mod service;
 
-use crate::service::GrpcWeb;
 use std::future::Future;
 use std::pin::Pin;
 use tonic::body::BoxBody;
